Fixed a bug on Campaign Groupings, added new campaign feature to add …

+++ adwords_dashboard/models.py
@@ -270,8 +270,8 @@ class Campaign(models.Model):
     account = models.ForeignKey(DependentAccount)
     campaign_id = models.CharField(max_length=255, default='None')
     campaign_name = models.CharField(max_length=255, default='None')
-    campaign_cost = models.BigIntegerField(default=0)
-    campaign_budget = models.IntegerField(default=0)
+    campaign_cost = models.FloatField(default=0)
+    campaign_budget = models.FloatField(default=0)
     groupped = models.BooleanField(default=False)
 
     @property
